refactor : Extract internal communication authentication in a reusable module
