+++ presto-internal-communication/src/main/java/com/facebook/presto/server/InternalAuthenticationManager.java
