+++ presto-internal-communication/src/main/java/com/facebook/presto/server/security/InternalAuthenticationFilter.java
