+++ presto-main/pom.xml
@@ -397,6 +397,12 @@
             <artifactId>json-smart</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>com.facebook.presto</groupId>
+            <artifactId>presto-internal-communication</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
         <!-- Testing dependencies -->
         <dependency>
             <groupId>com.facebook.presto</groupId>
